fix(phase-14): resolve 4 Codex-verified bugs + invitation_token orgId
- Fix org-context UUID-path fallback treating resourceId as orgId on nested routes
- Add request body extraction for organizationId (POST/PUT/PATCH fallback)
- Fix public directory allowlist mismatch (directory/public → directory/search)
- Fix event form BigInt serialization (API expects number, not BigInt)
- Add organizationId column to invitation_token table with FK + index
- 2069 tests passing, 0 failures

+++ apps/memberry/src/features/events/components/event-form.tsx
@@ -82,7 +82,7 @@ export function EventForm({ orgId, event, onSuccess, onCancel }: EventFormProps)
       startDate: new Date(form.startDate),
       endDate: new Date(form.endDate),
       location: form.location || undefined,
-      registrationFee: BigInt(Math.round(parseFloat(form.registrationFee || '0') * 100)),
+      registrationFee: Math.round(parseFloat(form.registrationFee || '0') * 100),
       capacity: form.capacity ? parseInt(form.capacity, 10) : undefined,
       creditBearing: false,
     }
